+++ lets/redirect/tcp/tls.py
@@ -1,4 +1,5 @@
 from lets.__module__ import Module, Mount, Container
+import argparse
 
 
 class TLS(Module):
@@ -6,7 +7,7 @@ class TLS(Module):
     Redirect TCP traffic over TLS.
     """
     images = [
-        "local/socat:1.0.0",      # Cross-platform
+        "local/socat:1.0.1",      # Cross-platform
         ]
 
     @classmethod
@@ -16,20 +17,65 @@ class TLS(Module):
         parser.add_argument("lport", type=int, help="listen on port")
         parser.add_argument("rhost", type=str, help="connect to host")
         parser.add_argument("rport", type=int, help="connect to port")
+        parser.add_argument("-k", "--key", type=argparse.FileType("rb"),
+            help="use TLS key")
+        parser.add_argument("-c", "--certificate", type=argparse.FileType("rb"),
+            help="use TLS certificate")
+        parser.add_argument("--ca", type=argparse.FileType("rb"),
+            help="use tls certificate authority to verify peer")
 
-    def handle(self, input, lhost, lport, rhost, rport):
+    def handle(self, input, lhost, lport, rhost, rport, key=None, certificate=None, ca=None):
 
-        # Construct command
-        cmd  = ""
-        cmd += " tcp-listen:%i,bind=%s,fork,reuseaddr" % (lport, lhost)
-        cmd += " openssl-connect:%s:%i,verify=0" % (rhost, rport)
+        with Mount("/data") as mount:
 
-        # Launch redirector
-        with Container.run("local/socat:1.0.0",
-            network="host",     # Use host network to allow local addresses
-            tty=True,
-            stdin_open=True,
-            command=cmd) as container:
+            # Construct command
+            cmd  = "socat"
+            cmd += " tcp-listen:%i,bind=%s,fork,reuseaddr" % (lport, lhost)
+            cmd += " openssl-connect:%s:%i" % (rhost, rport)
+            cmd += ",key=/data/key.pem,cert=/data/cert.pem"
 
-            self.log.info("Redirecting TCP(%s:%i) to TLS(%s:%i)", lhost, lport, rhost, rport)
-            container.interact()
+            # Write supplied TLS key/certificate
+            if key and certificate:
+
+                self.log.debug("Using key: %s", key.name)
+                with mount.open("key.pem", "wb") as f:
+                    f.write(key.read())
+
+                self.log.debug("Using certificate: %s", certificate.name)
+                with mount.open("cert.pem", "wb") as f:
+                    f.write(certificate.read())
+            
+            # Write generated TLS key/certificate
+            else:
+                if key:
+                    self.log.warning("Missing certificate, using self-signed")
+                if certificate:
+                    self.log.warning("Missing key, using self-signed")
+
+                self.log.debug("Generating self-signed certificate")
+                import lets.generate.tls.certificate
+                pem = lets.generate.tls.certificate()
+                with mount.open("key.pem", "wb") as f:
+                    f.write(pem)
+                with mount.open("cert.pem", "wb") as f:
+                    f.write(pem)
+
+            # Write certificate authority
+            if ca:
+                self.log.debug("Verifying peer with certificate: %s", ca.name)
+                with mount.open("ca.pem", "wb") as f:
+                    f.write(ca.read())
+                cmd += ",cafile=/data/ca.pem"
+            else:
+                cmd += ",verify=0"
+
+            # Launch redirector
+            with Container.run("local/socat:1.0.1",
+                network="host",     # Use host network to allow local addresses
+                tty=True,
+                stdin_open=True,
+                volumes=mount.volumes,
+                command=cmd) as container:
+
+                self.log.info("Redirecting TCP(%s:%i) to TLS(%s:%i)", lhost, lport, rhost, rport)
+                container.interact()